refactor: tighten coordinator package boundaries
